pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-fetch-metadata"
version = "0.1.0"
description = "Resource isolation policy for Django using Fetch Metadata request headers."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Luke Rogers", email = "luke@dmptr.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "django>=4.2",
]

[project.optional-dependencies]
toolbar = ["django-debug-toolbar>=4.0"]

[project.urls]
Homepage = "https://github.com/dmptrluke/django-fetch-metadata"

[tool.hatch.build.targets.wheel]
packages = ["fetch_metadata"]
exclude = ["fetch_metadata/tests"]

[tool.ruff]
target-version = "py310"
line-length = 120
extend-exclude = [
    "node_modules",
    "build",
]

[tool.ruff.lint]
select = [
    "E", "W",    # pycodestyle
    "F",         # pyflakes
    "I",         # isort
    "UP",        # pyupgrade
    "DJ",        # flake8-django
    "A",         # flake8-builtins
    "B",         # flake8-bugbear
    "C4",        # flake8-comprehensions
    "SIM",       # flake8-simplify
    "RUF",       # ruff-specific
    "PIE",       # flake8-pie
    "S",         # bandit/security
    "T20",       # flake8-print
    "T10",       # flake8-debugger
    "BLE",       # flake8-blind-except
    "LOG",       # flake8-logging
    "ISC",       # flake8-implicit-str-concat
    "FLY",       # flynt
    "FURB",      # refurb
    "RET",       # flake8-return
    "N",         # pep8-naming
    "PGH",       # pygrep-hooks
    "RSE",       # flake8-raise
    "C90",       # mccabe
]
ignore = [
    "RUF012",  # mutable class attrs (false positive for Django)
    "SIM103",  # return condition directly
    "SIM108",  # ternary operator
    "RET504",  # unnecessary assignment before return
    "RET505",  # unnecessary else after return
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["S"]
"**/conftest.py" = ["S"]
"fetch_metadata/test.py" = ["S101"]

[tool.ruff.format]
quote-style = "single"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.6",
]
test = [
    "pytest>=9.0",
    "pytest-django>=4.11",
    "django-debug-toolbar>=4.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "fetch_metadata.tests.settings"

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "django", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
"django" = ["django"]
